requirements.txt
torch>=2.10.0
torchvision>=0.25.0
tensorboard>=2.20.0
Pillow>=12.1.0
numpy>=2.4.0
scikit-learn>=1.8.0
tqdm>=4.67.3
matplotlib>=3.10.8
gradio>=5.0.0
